selenium/requirements-windows.txt
selenium==4.16.0
webdriver-manager==4.0.1
pydantic>=2.3.0,<3
selenium-wire==5.1.0
pywin32==306
colorlog==6.8.0
pillow==10.2.0
pyzbar==0.1.9
pydantic-settings==2.1.0
pytest-lazy-fixture==0.6.3